Improve XC flying potential index
Previously, the computation was too penalized when there was a little bit of clouds.
The new computation uses a logistic function, just like it was done in soaringmeteo v1. I believe the new results are a bit better, but we experience will tell us if we need to tweak again the formula manually...