Don’t download GRIB files for old-soarGFS anymore
The implementation was motivated by the desire to download the GFS data once and use it for both soarV2 and soarGFS. However, it didn’t work in practice because the GRIB files downloaded here are significantly larger than the ones downloaded by old soarGFS, and then old soarGFS takes too long time to analyze them and eventually fails.

In this commit, we don’t try to download the GRIB files anymore in way that they can be reused by old soarGFS. The consequence is that both soarV2 and old soarGFS will download data from the NOAA servers, so we need to be careful to not do it at the same time to avoid hitting the rate limit (120 hits/min).