Show numerical values for wind velocity
Introduce a new setting that, when enabled, displays the wind with a simple arrow and the wind speed (as a number), as opposed to the default wind barbells.

This commit implements the feature on meteograms and sounding diagrams (but not on the map overlay).

The setting is automatically saved and restored across user sessions.