Support breaking changes in the forecast data format
Currently, changing the format of the forecast data produced by the server is delicate to handle: the frontend still shows the older (incompatible) forecast runs, and more importantly there is a short period of time where the frontend is not able to consume the data produced by the server.

We now indicate the version of the format of the data produced by the backend and consumed by the frontend so that whenever we need to change the format of the data produced by the backend, we can simply delay the deployment of the frontend and only when we will deploy the updated frontend then the new forecast data will be consumed (in the meantime, the frontend will still work with the older forecast data).