Use OpenLayers as a map engine
This is the first chunk of work to support [WRF forecasts in soarV2](https://github.com/orgs/soaringmeteo/projects/1). Most of the system has been rewritten to support processing up to 500 million data points in total (which represents an increase of about 10 times compared to the current state).

On the frontend, we replace Leaflet with OpenLayers to render the map and its overlays. See the corresponding decision record in the `docs/decisions/` directory for more information on the decision drivers and the considered options.

On the backend, we cannot anymore afford to process all the data in the memory, and we now use an intermediate storage on the disk. We achieve that with the H2 embedded database. See the corresponding decision record for more information.

We also change the format of the data produced by the backend: we now produce raster images and vector tiles in addition to JSON documents.

For the end-users, the main change is that they now need to select which zone to display within the GFS domain. By default, we display the Europe only.