Generate frontend assets from WRF forecast results
This is the second chunk of work aiming at displaying [WRF forecasts in soarV2](https://github.com/orgs/soaringmeteo/projects/1).

We modify the backend to produce raster images and vector tiles from WRF forecast results. The generation of JSON data (required to generate meteograms and sounding diagrams) will be achieved in a separate PR.

We change the structure of the project to have three modules: common, gfs, and wrf. The modules gfs and wrf depend on common, which provides the main pieces of logic (generation of raster images and vector tiles, definition of data format, etc.).

We applied just enough changes to the frontend to display the new overlays, but more changes will be necessary in the long run to support meteograms and to improve the overall user experience.