Make colors in legends look a bit more like the colors on the map
Because the map overlays are a bit transparent, the colors seen in the overlay legend (bottom right of the screen) can look pretty different from those seen on the map overlays.

We apply some transparency to the colors shown in the legend as well to make them look more similar to the colors in the overlays.