Simplify the structure of forecasts metadata.
Instead of storing each forecast metadata in their own file, store an array of “active” forecasts in the file `forecast.json`.

This simplifies and speeds up the loading logic on the frontend, and also fix reliability issue when computing the initialization date of a forecast based on the assets path.

Another benefit is that we don’t need to add the resolution to the zone labels since all the zones of a model have the same resolution.