Go back to a single soarWRF model
Instead of showing WRF6 as a different model from WRF2, keep them under the same name. Users won’t have to switch between models, and they will automatically discover that we now compute forecasts for the next 4 days. When they select the run for d+3 or d+4, we automatically fall back to the 6-km resolution grid.